backend/requirements.txt
Django>=4.2,<5.0
djangorestframework>=3.14,<3.16
psycopg2-binary>=2.9,<2.10
django-cors-headers>=3.0,<4.4
djangorestframework-simplejwt>=5.0,<5.4
drf-spectacular>=0.26,<0.28
